Simplified Mqtt3PubAck/Rec/Rel/CompEncoder
