+++ src/main/java/org/mqttbee/mqtt/codec/encoder/mqtt3/Mqtt3PubRelEncoder.java
@@ -17,10 +17,8 @@
 
 package org.mqttbee.mqtt.codec.encoder.mqtt3;
 
-import io.netty.buffer.ByteBuf;
-import org.mqttbee.annotations.NotNull;
 import org.mqttbee.api.mqtt.mqtt3.message.Mqtt3MessageType;
-import org.mqttbee.mqtt.codec.encoder.mqtt3.Mqtt3MessageEncoder.Mqtt3MessageFixedSizeEncoder;
+import org.mqttbee.mqtt.codec.encoder.mqtt3.Mqtt3MessageEncoder.Mqtt3MessageWithIdEncoder;
 import org.mqttbee.mqtt.message.publish.pubrel.MqttPubRel;
 
 import javax.inject.Inject;
@@ -31,34 +29,17 @@ import javax.inject.Singleton;
  * @author Silvio Giebl
  */
 @Singleton
-public class Mqtt3PubRelEncoder extends Mqtt3MessageFixedSizeEncoder<MqttPubRel> {
+public class Mqtt3PubRelEncoder extends Mqtt3MessageWithIdEncoder<MqttPubRel> {
 
     private static final int FIXED_HEADER = (Mqtt3MessageType.PUBREL.getCode() << 4) | 0b0010;
-    private static final int REMAINING_LENGTH = 2;
-    private static final int ENCODED_LENGTH = 2 + REMAINING_LENGTH;
 
     @Inject
     Mqtt3PubRelEncoder() {
     }
 
     @Override
-    int encodedLength() {
-        return ENCODED_LENGTH;
-    }
-
-    @Override
-    public void encode(@NotNull final MqttPubRel message, @NotNull final ByteBuf out) {
-        encodeFixedHeader(out);
-        encodeVariableHeader(message, out);
-    }
-
-    private void encodeFixedHeader(@NotNull final ByteBuf out) {
-        out.writeByte(FIXED_HEADER);
-        out.writeByte(REMAINING_LENGTH);
-    }
-
-    private void encodeVariableHeader(@NotNull final MqttPubRel message, @NotNull final ByteBuf out) {
-        out.writeShort(message.getPacketIdentifier());
+    int getFixedHeader() {
+        return FIXED_HEADER;
     }
 
 }
